Move filtering to ModelManager
