--- src/main/java/seedu/address/parser/expr/Expr.java
@@ -1,8 +1,0 @@
-package seedu.address.parser.expr;
-
-import seedu.address.model.person.ReadOnlyPerson;
-
-public interface Expr {
-    boolean satisfies(ReadOnlyPerson person);
-    String toString();
-}
