--- src/main/java/seedu/address/parser/qualifier/Qualifier.java
@@ -1,9 +1,0 @@
-package seedu.address.parser.qualifier;
-
-
-import seedu.address.model.person.ReadOnlyPerson;
-
-public interface Qualifier {
-    boolean run(ReadOnlyPerson person);
-    String toString();
-}
