--- src/main/java/seedu/address/parser/qualifier/TrueQualifier.java
@@ -1,19 +1,0 @@
-package seedu.address.parser.qualifier;
-
-import seedu.address.model.person.ReadOnlyPerson;
-
-public class TrueQualifier implements Qualifier {
-
-    public TrueQualifier() {
-    }
-
-    @Override
-    public boolean run(ReadOnlyPerson person) {
-        return true;
-    }
-
-    @Override
-    public String toString() {
-        return "TRUE";
-    }
-}
